frontend/requirements.txt
confluent_kafka==2.10.0
pandas==2.2.3
plotly==6.1.2
python_dateutil==2.9.0.post0
streamlit~=1.34.0
pillow~=9.1.1
streamlit_js_eval==0.1.7
matplotlib==3.10.3
httpx==0.28.1